pyproject.toml
[tool.poetry]
name = "chronoseek-subnet"
version = "0.1.0"
description = "ChronoSeek (Semantic Video Moment Retrieval) on Bittensor"
authors = ["CreativeEssence <dalyconnor528@gmail.com>"]
readme = "README.md"
packages = [
    { include = "chronoseek" }
]

[tool.poetry.dependencies]
python = ">=3.12, <3.15"
bittensor = ">=10.0.0"
bittensor-cli = ">=9.18.0"
bittensor-wallet = ">=4.0.0"
click = ">=8.0.0"
fastapi = ">=0.109.0"
uvicorn = ">=0.27.0"
httpx = ">=0.26.0"
pydantic = ">=2.0.0"
numpy = ">=1.24.0"
torch = "^2.10.0"
transformers = "^5.2.0"
opencv-python = "^4.13.0.92"
ftfy = "^6.3.1"
regex = "^2026.2.28"
requests = "^2.32.5"
pillow = "^12.1.1"
python-dotenv = "^1.0.0"
jsonschema = "^4.25.1"
huggingface-hub = ">=1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
